requirements.txt
fastapi
pyjwt
passlib
types-passlib
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv